requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
sqlmodel
sqlalchemy
pydantic
typing-extensions
boto3
psycopg2-binary
